added ARKLogTypes.h
