+++ Aardvark/ARKLogTypes.h
@@ -1,0 +1,32 @@
+//
+//  ARKLogTypes.h
+//  Aardvark
+//
+//  Created by Evan Kimia on 10/23/15.
+//  Copyright 2015 Square, Inc.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+#import <UIKit/UIKit.h>
+
+typedef NS_ENUM(NSUInteger, ARKLogType) {
+    /// Default log type.
+    ARKLogTypeDefault,
+    /// Marks the beginning or end of a task.
+    ARKLogTypeSeparator,
+    /// Marks that the log represents an error.
+    ARKLogTypeError,
+    /// Marks a log that has a screenshot attached.
+    ARKLogTypeScreenshot,
+};
